requirements.txt
pyinotify
loguru
python-nmap


